Criado o controller da Açõo de extenção

+++ aplicacao/gpa-extensao/src/main/java/ufc/quixada/npi/gpa/repository/AcaoExtensaoRepository.java
@@ -11,7 +11,7 @@ import ufc.quixada.npi.gpa.model.AcaoExtensao;
 @Repository
 public interface AcaoExtensaoRepository extends CrudRepository<AcaoExtensao, Integer> {
 	
-	AcaoExtensao getById(Long id);
+	AcaoExtensao getById(int id);
 	
 	
 }
