+++ aplicacao/gpa-extensao/src/main/java/ufc/quixada/npi/gpa/service/AcaoExtensaoService.java
@@ -5,6 +5,6 @@ package ufc.quixada.npi.gpa.service;
 import ufc.quixada.npi.gpa.model.AcaoExtensao;
 
 public interface AcaoExtensaoService {
-	AcaoExtensao getById(Long id);
+	AcaoExtensao getById(int id);
 	
 }
