+++ aplicacao/gpa-extensao/src/main/java/ufc/quixada/npi/gpa/service/impl/AcaoExtensaoServiceImpl.java
@@ -5,7 +5,6 @@ package ufc.quixada.npi.gpa.service.impl;
 import javax.inject.Named;
 
 import org.springframework.beans.factory.annotation.Autowired;
-
 import ufc.quixada.npi.gpa.model.AcaoExtensao;
 import ufc.quixada.npi.gpa.repository.AcaoExtensaoRepository;
 import ufc.quixada.npi.gpa.service.AcaoExtensaoService;
@@ -17,7 +16,7 @@ public class AcaoExtensaoServiceImpl implements AcaoExtensaoService{
 	AcaoExtensaoRepository acaoRepository;
 	
 	@Override
-	public AcaoExtensao getById(Long id) {
+	public AcaoExtensao getById(int id) {
 		return acaoRepository.getById(id);
 	}
